Add biases & more activations
Implemented biases in the model
Added linear and leaky ReLU activations

+++ src/main/java/core/Matrix.java
@@ -63,7 +63,7 @@ public class Matrix {
         }
     }
 
-    public Matrix mult(int x) {
+    public Matrix mult(double x) {
         Matrix out = clone();
         for (int i = 0; i < rows; i++) {
             for (int j = 0; j < cols; j++) {
